+++ ts/features/euCovidCert/screens/EuCovidCertLoadingScreen.tsx
@@ -25,7 +25,10 @@ const euActivityIndicator = (
 const EuCovidCertLoadingScreen = (_: Props): React.ReactElement => (
   <BaseEuCovidCertificateLayout
     testID={"EuCovidCertLoadingScreen"}
-    /* don't show header during the loading phase since it could be replaced with the one with certificate data */
+    /*
+    don't show header during the loading phase since it could be replaced with the one with certificate data
+    causing an annoying 'glitch' with the headers replacement
+     */
     header={<></>}
     content={
       <View>
